pyproject.toml
[tool.poetry]
name = "doopliss"
version = "0.1.0"
description = ""
authors = ["Nora Hanegan <NoraH@JRISchools.ORG>"]

[tool.poetry.dependencies]
python = "^3.8"
google-auth-oauthlib = "^0.4.2"
opencv-python = "^4.4.0"
python-dotenv = "^0.15.0"
py-cord = "^1.7.3"
google-api-python-client = "^1.12.8"
flask = "^1.1.2"
python-box = "^5.2.0"
dadjokes = "^1.3.2"
discord = "^1.7.3"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
